move func `condition_shotnum` from `hdfreadcontrol.py` to `helpers.py`; did the same with covering tests
